:stopwatch: measure processing time ratios
We are interested in tracking how long diarization and transcription
take relative to the length of the recorded meeting. For example,
If a ten-minute meeting takes 5 minutes to diarize, then the ratio
is 0.5.

We like to record the information so that we can later collect it
and fine-tune the numbers for different machines. For example, we
expect our vultr.com machine (test.diarizer.com) to be much slower
than our ESXi machine (diarizer.com)

Notes
- test a floating point number (an integer-only test won't uncover
  bugs from division + conversion(int64, time.Duration, float64)