Remove time zone ("PDT") :clock6: information from test
This corrects a bug caused by time zone information being lost
when the JSON file is written out & read back in. When
we output our time to JSON, we elide the time zone (i.e.
we lose the "PDT", (though we maintain the offset (-25200)))

This causes a failure when running on machines whose time zone
is not set to PDT, machines such as our CI server, which displays
time in UTC.

This bug was particularly difficult to troubleshoot,
for the failure output (see below) did not included time zone,
thus the two items appeared identical even though the tests
claimed they were different.

fixes: <https://ci.nono.io/teams/main/pipelines/BlabberTabber/jobs/unit/builds/60>
```
• Failure [0.001 seconds]
TimesAndSize
/tmp/build/b4e0c68a/src/timesandsize/timesandsize_test.go:49
  .WriteTimesAndSize
  /tmp/build/b4e0c68a/src/timesandsize/timesandsize_test.go:48
    When a writer is called on a valid TimesAndSize struct
    /tmp/build/b4e0c68a/src/timesandsize/timesandsize_test.go:47
      writes out the JSON struct to the writer [It]
      /tmp/build/b4e0c68a/src/timesandsize/timesandsize_test.go:46

      Expected
          <timesandsize.TimesAndSize>: {
              WaveFileSizeInBytes: 256,
              Diarizer: "blah",
              Transcriber: "blech",
              DiarizationProcessingRatio: 0,
              TranscriptionProcessingRatio: 0,
              EstimatedDiarizationFinishTime: 2017-09-04T20:49:14-07:00,
              EstimatedTranscriptionFinishTime: 2017-09-04T20:49:15-07:00,
          }
      to equal
          <timesandsize.TimesAndSize>: {
              WaveFileSizeInBytes: 256,
              Diarizer: "blah",
              Transcriber: "blech",
              DiarizationProcessingRatio: 0,
              TranscriptionProcessingRatio: 0,
              EstimatedDiarizationFinishTime: 2017-09-04T20:49:14-07:00,
              EstimatedTranscriptionFinishTime: 2017-09-04T20:49:15-07:00,
          }
```