The CMU transcriber is officially "CMU Sphinx 4"
But we also allow "CMUSphinx4" and "CMU Sphinx4" for a couple of months
until the Android clients have been updated to use the new transcriber.

The Android client now uses "CMU Sphinx 4" because that matches the text
field, and it's easier and simpler to have one string to be both the
label (text field in the menu) and the requested diarizer.

fixes:
```
I couldn't find Transcriber["CMU Sphinx4"]!
```